int typedef

+++ src/fov.c
@@ -1,40 +1,39 @@
-#include "base.h"
 #include "map.h"
 
-static int multipliers[4][8] = {
+static i8 multipliers[4][8] = {
   {1, 0, 0, -1, -1, 0, 0, 1},
   {0, 1, -1, 0, 0, -1, 1, 0},
   {0, 1, 1, 0, 0, -1, -1, 0},
   {1, 0, 0, 1, -1, 0, 0, -1}
 };
 
-void cast_light(uint8_t x, uint8_t y, uint8_t radius, uint8_t row, float start_slope,
-    float end_slope, int8_t xx, int8_t xy, int8_t yx, int8_t yy) {
+void cast_light(u8 x, u8 y, u8 radius, u8 row, float start_slope,
+    float end_slope, i8 xx, i8 xy, i8 yx, i8 yy) {
 
   if (start_slope < end_slope) return;
 
   float next_start_slope = start_slope;
-  for (uint8_t i = row; i <= radius; i++) {
+  for (u8 i = row; i <= radius; i++) {
     bool blocked = false;
-    for (int8_t dx = -i, dy = -i; dx <= 0; dx++) {
+    for (i8 dx = -i, dy = -i; dx <= 0; dx++) {
       float l_slope = (dx - 0.5) / (dy + 0.5);
       float r_slope = (dx + 0.5) / (dy - 0.5);
 
       if (start_slope < r_slope) continue;
       else if (end_slope > l_slope) break;
 
-      int8_t sax = dx * xx + dy * xy;
-      int8_t say = dx * yx + dy * yy;
+      i8 sax = dx * xx + dy * xy;
+      i8 say = dx * yx + dy * yy;
 
-      if ((sax < 0 && (int8_t) abs(sax) > x) || (say < 0 && (int8_t) abs(say) > y)) continue;
+      if ((sax < 0 && (i8) abs(sax) > x) || (say < 0 && (i8) abs(say) > y)) continue;
 
-      uint8_t ax = x + sax;
-      uint8_t ay = y + say;
+      u8 ax = x + sax;
+      u8 ay = y + say;
 
       if (ax >= WIDTH || ay >= HEIGHT) continue;
 
-      uint8_t radius2 = radius * radius;
-      if ((int8_t) (dx * dx + dy * dy) < radius2) {
+      u8 radius2 = radius * radius;
+      if ((i8) (dx * dx + dy * dy) < radius2) {
         m->gr[ay * WIDTH + ax]->vb = 1;
         m->gr[ay * WIDTH + ax]->vt = 1;
         m->gr[ay * WIDTH + ax]->co = 2;
@@ -59,9 +58,9 @@ void cast_light(uint8_t x, uint8_t y, uint8_t radius, uint8_t row, float start_s
   }
 }
 
-void fov(uint8_t x, uint8_t y, uint8_t radius) {
+void fov(u8 x, u8 y, u8 radius) {
   reset_grid_visibility();
-  for (uint8_t i = 0; i < 8; i++) {
+  for (u8 i = 0; i < 8; i++) {
     cast_light(x, y, radius, 1, 1.0, 0.0, multipliers[0][i],
         multipliers[1][i], multipliers[2][i], multipliers[3][i]);
   }
